requirements.txt
git+https://github.com/boettiger-lab/cng-python
altair==5.3.0
duckdb==1.2.2
duckdb_engine== 0.15.0
geoarrow-types==0.2.0
geoarrow-pandas==0.1.1
ibis-framework[duckdb]==10.3.1
leafmap==0.53.3
minio==7.2.15
maplibre==0.3.3
pandas==2.2.3
shiny==0.10.2
SQLAlchemy==2.0.35
streamlit==1.50.0